griffin: use more conservative VM parameters
The maximum heap size for each virtual machine is 1GB. This is too big
and puts too much pressure on both the garbage collector when it tries
to perform collections in older generations. The symptoms are simple,
some applications will allocate objects that live long enough to be
promoted to older generations, escaping the efficient CMS collections.
Eventually, a stop-the-world collection will go through this large block
of memory and cause applications to pause for too long.

Additionally, the out of memory killer follows a LRU approach where it
tries to kill older applciations to make room for new ones. The problem
arises when the least recently used applications don't have large heaps
but the most recent ones do. When this happens the system halts to a
crawl while older apps and services are killed and respawned in a never
ending cycle of jankiness.

These values were obtained experimentally but actually match the
defaults for devices with >2gb of ram and xhdpi.

Signed-off-by: Erfan Abdi <erfangplus@gmail.com>

# Conflicts:
#	device.mk